refactor: 중복되는 내부 클래스 dto 외부로 추출
