+++ src/main/java/com/moa/dto/recruit/RecruitmentsInfo.java
@@ -3,7 +3,6 @@ package com.moa.dto.recruit;
 import com.moa.domain.recruit.Recruitment;
 import lombok.Getter;
 
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -21,29 +20,5 @@ public class RecruitmentsInfo {
             infos.add(RecruitmentInfo.of(recruitment, 0));
         }
         return infos;
-    }
-    
-    public record RecruitmentInfo(
-            Long id,
-            String title,
-            String author,
-            String category,
-            String recruitStatus,
-            String createAt,
-            String profileImage,
-            int replyCount
-    ) {
-
-        public static RecruitmentInfo of(Recruitment recruitment, int replyCount) {
-            return new RecruitmentInfo(recruitment.getId(),
-                    recruitment.getPost().getTitle(),
-                    recruitment.getUser().getName(),
-                    recruitment.getCategory().getName(),
-                    recruitment.getStatus().getStatus(),
-                    recruitment.getCreatedDate().format(DateTimeFormatter.ofPattern("yyyy-MM-dd")),
-                    "profileImageLink",
-                    replyCount
-            );
-        }
     }
 }
